update bootstrap dependency in package.json as suggested by Github
Reason:
CVE-2018-14041 More information

moderate severity
Vulnerable versions: < 4.1.2
Patched version: 4.1.2
In Bootstrap before 4.1.2, XSS is possible in the data-target property of scrollspy. This is similar to CVE-2018-14042.

Remediation

Upgrade bootstrap to version 4.1.2 or later. For example:

"dependencies": {
  "bootstrap": ">=4.1.2"
}
or…
"devDependencies": {
  "bootstrap": ">=4.1.2"
}
Always verify the validity and compatibility of suggestions with your codebase.